dom/media/mediasource/test/test_AudioChange_mp4.html

<!DOCTYPE HTML>
<html>
<head>
  <title>MSE: basic functionality</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="mediasource.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

// This test checks loading a stereo segment, followed by a 5.1 segment plays without error.

runWithMSE(function(ms, el) {
  el.controls = true;
  once(ms, "sourceopen").then(function() {
    // Log events for debugging.
    const events = [ "suspend", "play", "canplay", "canplaythrough", "loadstart", "loadedmetadata",
                     "loadeddata", "playing", "ended", "error", "stalled", "emptied", "abort",
                     "waiting", "pause", "durationchange", "seeking", "seeked" ];
    function logEvent(e) {
      info("got " + e.type + " event");
    }
    events.forEach(function(e) {
      el.addEventListener(e, logEvent);
    });

    ok(true, "Receive a sourceopen event");
    const audiosb = ms.addSourceBuffer("audio/mp4");
    el.addEventListener("error", function(e) {
      ok(false, "should not fire '" + e.type + "' event");
      SimpleTest.finish();
    });
    is(el.readyState, el.HAVE_NOTHING, "readyState is HAVE_NOTHING");
    const loadedmetadataPromises = [];
    loadedmetadataPromises.push(fetchAndLoad(audiosb, "aac20-48000-64000-", [ "init" ], ".mp4"));
    loadedmetadataPromises.push(once(el, "loadedmetadata"));
    Promise.all(loadedmetadataPromises)
    .then(function() {
      ok(true, "got loadedmetadata event");
      const canplayPromises = [];
      canplayPromises.push(once(el, "loadeddata"));
      canplayPromises.push(once(el, "canplay"));
      canplayPromises.push(fetchAndLoad(audiosb, "aac20-48000-64000-", [ "1" ], ".m4s"));
      return Promise.all(canplayPromises);
    })
    .then(function() {
      ok(true, "got canplay event");
      el.play();
      return fetchAndLoad(audiosb, "aac51-48000-128000-", [ "init" ], ".mp4");
    })
    .then(fetchAndLoad.bind(null, audiosb, "aac51-48000-128000-", [ "2" ], ".m4s"))
    .then(function() {
      ms.endOfStream();
      return once(el, "ended");
    })
    .then(function() {
      ok(el.currentTime >= 6, "played to the end");
      SimpleTest.finish();
    });
  });
});

</script>
</pre>
</body>
</html>
